Move FillBlockAfterExecutionAsync into BlockExecutingService
